move otu plan functions under otu.modify
